Accounted for population growth between steady state and first year of transition.
ModelTester displayed numerical deviations for the steady state, which I believe to be due to the substitution of the variable pgr by its formula in the calculation of bequests. Modeltester also shows small deviations (around 0.78%, which is similar to the population growth rate) for most cases, except for 19% caps_foreign in the open cases and 27% expsubs in the closed baseline case.