Computed the mass of (working-age) people not working at all and that of people with zero consumption.
Note: I consider zero hours worked and zero consumption to be anything smaller than 1e-10. When I change that value to 0, nothing changes for consumption, that is, the mass remains equal to zero. Meanwhile, the mass of people working exactly zero hours drops to zero. This means there are households working an arbitrarily small number of hours, a case for discretizing labor supply.