Added modelunit_dollar to the creation of assets and average earnings grids (instead of hard coding it).
Changed parameters in the average earnings grid such that its values are close to the ones before the change. Changed construction of assets grid such that only the last gridpoint is useless (with a mass of households around 1e-14). ModelTester shows small deviations for most variables, except for around 1% deviations for bequests, around 300% deviation for caps_foreign and cits_foreign (open), and around 7% deviations for expsubs (closed).