Created a variable for capital gains/losses instead of inflating the capital grid. Applied taucap to newly priced asset holdings in solve_cohort, consistent to what we were doing with the kpricescale method.
ModelTester does not match for counter cases with, mostly, tiny deviations due to discretization errors. The largest deviations (6% in open_counter caps_foreign/debts_foreign/cits_foreign and 13% in closed_counter expsubs) also arise from discretization errors, but magnified due to magnitude of these variables.

TBD1: Repeat the last diff element in expsubs, instead of assuming it to be zero.
TBD2: Rethink the unit  in which caprates is denominated (we may want to divide by qtobin).