Corrected bequests redistribution in the first transition year.
Not dividing by population in the first period was correct when pop=1 in the first transition year, that is, not accounting for population growth from steady state to transition.
ModelTester full match for steady state. Deviations around or below 5%, except for foreign capital (180%) in the open economies, expsubs (200%) in the closed economies, and cits (20%) in the open economies.