Replaced last period expsubs by a better approximation (considering the economy converges towards a new steady state, using t-1 value is better than 0).
ModelTester full match for all economies, except for open_counter. This is a PROBLEM because we should also observe deviations for open_base, as we are directly changing expsubs. For closed economies, having no deviations is ok because diff(caps) assumes negative values and expsubs takes max between that and zero. Deviations in open_counter are as expected: 40% deviation in expsubs, 7% in cits and cits_domestic, 2% in cits_foreign/caprevs/caps_foreign, and less than 1% in other variables.

TBD: investigate why open_base does not update expsubs values (it's always zero).