Introduced a guess for expsubs.
ModelTester results: numerical deviations for steady state and closed economies. For open_base: 51% caps_foreign and cits_foreign, 0.01% debts foreign and expsubs, numerical deviations for all others. Open_counter: similar to open_base.

NOTE: the size of numerical deviations varies substantially when we change the constraint on number of iterations. The deviations discussed above are for at least 2 iterations. For at least 3 iterations, the change is tiny: from 51% to 52%. But if we remove that condition, deviations jump to 111% although expsubs do not change.