Replaced investment net of depreciation by gross investment.
ModelTester displays substantial deviations. Only deviation for steady state is 167% in investment. For open_base, largest deviations are ~600% for cits and cits_domestic, followed by 314% for caprevs. For open_counter, 840% caps/cits_foreign and 784% for caprevs. For closed_base, 1048% for cits and cits_domestic and 303% caprevs. For closed_counter, 523% cits and cits_domestic and 461% caprevs. Those are only the largest deviations.