Used before re-pricing/beginning-of-period variables to calculate domestic capital in the open economy.
Now the difference between the capital inherited from steady state and that of the first period of open economy transition is in order of 1e-05, compatible with our tolerances. The difference was 0.016 using the previous caps_domestic equation.
TBD: Investigate why the gap between capital inherited from steady state and that of the first period of closed economy transition does not shrink when changing its caps_domestic calculation.

ModelTester full match for steady state and closed economies.
For open economies only deviations are 0.02% caps_domestic and 0.89% caps_foreign.